+++ web/js/telemetry/MachineDatastoreHooks.tsx
@@ -1,7 +1,7 @@
 import {MachineDatastore} from "./MachineDatastores";
 import {useFirestore} from "../../../apps/repository/js/FirestoreProvider";
 import {MachineIDs} from "polar-shared/src/util/MachineIDs";
-import * as firebase from 'firebase/app';
+import firebase from 'firebase/app'
 import {OnErrorCallback} from "polar-shared/src/util/Snapshots";
 import DocumentSnapshot = firebase.firestore.DocumentSnapshot;
 import {useSnapshots} from "../ui/data_loader/UseSnapshotSubscriber";